test(structure): split tests into Unit and Feature suites
