+++ .gitignore
@@ -2,3 +2,4 @@
 /vendor
 /.phpunit.cache
 composer.lock
+.phpunit.result.cache
