+++ tests/Feature/LogApiRequestMiddlewareTest.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace CodeTech\ApiLogs\Tests;
+namespace CodeTech\ApiLogs\Tests\Feature;
 
 use CodeTech\ApiLogs\Http\Middleware\LogApiRequest;
 use CodeTech\ApiLogs\Models\ApiLog;
 use CodeTech\ApiLogs\Tests\Fixtures\User;
+use CodeTech\ApiLogs\Tests\TestCase;
 
 class LogApiRequestMiddlewareTest extends TestCase
 {
