+++ tests/Unit/RedactorTest.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace CodeTech\ApiLogs\Tests;
+namespace CodeTech\ApiLogs\Tests\Unit;
 
 use CodeTech\ApiLogs\Support\Redactor;
-use PHPUnit\Framework\TestCase;
+use PHPUnit\Framework\TestCase as PHPUnitTestCase;
 
-class RedactorTest extends TestCase
+class RedactorTest extends PHPUnitTestCase
 {
     public function test_redacts_matching_keys_recursively(): void
     {
